+++ src/main/java/com/github/bgomar/consolelogger/toolwindow/setup/PropertiesConsoleLoggerToolSetup.java
@@ -19,6 +19,8 @@ public class PropertiesConsoleLoggerToolSetup  extends AbstractToolSetup impleme
     private static JTextField propertiesConsoleLoggerTextField8 = new JTextField();
     private static JTextField propertiesConsoleLoggerTextField9 = new JTextField();
     private static JButton propertiesConsoleLoggerSaveButton = new JButton();
+    private static JButton propertiesConsoleLoggerLoad2Button = new JButton();
+    private static JButton propertiesConsoleLoggerLoad1Button = new JButton();
     private static JButton propertiesConsoleLoggerCancelButton = new JButton();
 
     private static JButton propertiesConsoleLoggerDefaultButton1 = new JButton();
@@ -32,6 +34,8 @@ public class PropertiesConsoleLoggerToolSetup  extends AbstractToolSetup impleme
     private static JButton propertiesConsoleLoggerDefaultButton9 = new JButton();
 
     public static DialogPanel ui = new DialogPanel();
+
+    public int preset = 0;
 
     public PropertiesConsoleLoggerToolSetup() {
         setup();
@@ -48,6 +52,8 @@ public class PropertiesConsoleLoggerToolSetup  extends AbstractToolSetup impleme
             JTextField propertiesConsoleLoggerTextField8,
             JTextField propertiesConsoleLoggerTextField9,
             JButton propertiesConsoleLoggerSaveButton,
+            JButton propertiesConsoleLoggerLoad2Button,
+            JButton propertiesConsoleLoggerLoad1Button,
             JButton propertiesConsoleLoggerCancelButton,
             JButton propertiesConsoleLoggerDefaultButton1,
             JButton propertiesConsoleLoggerDefaultButton2,
@@ -69,8 +75,12 @@ public class PropertiesConsoleLoggerToolSetup  extends AbstractToolSetup impleme
         PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField7 = propertiesConsoleLoggerTextField7;
         PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField8 = propertiesConsoleLoggerTextField8;
         PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField9 = propertiesConsoleLoggerTextField9;
+
         PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerSaveButton = propertiesConsoleLoggerSaveButton;
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerLoad2Button = propertiesConsoleLoggerLoad2Button;
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerLoad1Button = propertiesConsoleLoggerLoad1Button;
         PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerCancelButton = propertiesConsoleLoggerCancelButton;
+
         PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton1 = propertiesConsoleLoggerDefaultButton1;
         PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton2 = propertiesConsoleLoggerDefaultButton2;
         PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton3 = propertiesConsoleLoggerDefaultButton3;
@@ -85,45 +95,111 @@ public class PropertiesConsoleLoggerToolSetup  extends AbstractToolSetup impleme
 
 
     public void setup() {
-        loadAllLoggers();
+        load1Loggers();
 
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton1.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField1.setText(ConsoleLoggerSettings.DEFAULT_PATTERN_1));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton2.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField2.setText(ConsoleLoggerSettings.DEFAULT_PATTERN_2));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton3.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField3.setText(ConsoleLoggerSettings.DEFAULT_PATTERN_3));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton4.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField4.setText(ConsoleLoggerSettings.DEFAULT_PATTERN_4));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton5.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField5.setText(ConsoleLoggerSettings.DEFAULT_PATTERN_5));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton6.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField6.setText(ConsoleLoggerSettings.DEFAULT_PATTERN_6));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton7.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField7.setText(ConsoleLoggerSettings.DEFAULT_PATTERN_7));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton8.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField8.setText(ConsoleLoggerSettings.DEFAULT_PATTERN_8));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton9.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField9.setText(ConsoleLoggerSettings.DEFAULT_PATTERN_9));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton1.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField1.setText((preset == 1) ? ConsoleLoggerSettings.DEFAULT_PATTERN_1 : ConsoleLoggerSettings.DEFAULT_PATTERN_10));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton2.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField2.setText((preset == 1) ? ConsoleLoggerSettings.DEFAULT_PATTERN_2 : ConsoleLoggerSettings.DEFAULT_PATTERN_11));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton3.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField3.setText((preset == 1) ? ConsoleLoggerSettings.DEFAULT_PATTERN_3 : ConsoleLoggerSettings.DEFAULT_PATTERN_12));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton4.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField4.setText((preset == 1) ? ConsoleLoggerSettings.DEFAULT_PATTERN_4 : ConsoleLoggerSettings.DEFAULT_PATTERN_13));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton5.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField5.setText((preset == 1) ? ConsoleLoggerSettings.DEFAULT_PATTERN_5 : ConsoleLoggerSettings.DEFAULT_PATTERN_14));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton6.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField6.setText((preset == 1) ? ConsoleLoggerSettings.DEFAULT_PATTERN_6 : ConsoleLoggerSettings.DEFAULT_PATTERN_15));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton7.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField7.setText((preset == 1) ? ConsoleLoggerSettings.DEFAULT_PATTERN_7 : ConsoleLoggerSettings.DEFAULT_PATTERN_16));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton8.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField8.setText((preset == 1) ? ConsoleLoggerSettings.DEFAULT_PATTERN_8 : ConsoleLoggerSettings.DEFAULT_PATTERN_17));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerDefaultButton9.addActionListener(e -> PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField9.setText((preset == 1) ? ConsoleLoggerSettings.DEFAULT_PATTERN_9 : ConsoleLoggerSettings.DEFAULT_PATTERN_18));
 
         PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerCancelButton.addActionListener(e -> {
-            loadAllLoggers();
+            if (preset == 1) {
+                load1Loggers();
+                saveActiveLoggers();
+            } else if (preset == 2)  {
+                load2Loggers();
+                saveActiveLoggers();
+            }
         });
+
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerLoad1Button.addActionListener(e -> {
+            load1Loggers();
+            saveActiveLoggers();
+        });
+
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerLoad2Button.addActionListener(e -> {
+            load2Loggers();
+            saveActiveLoggers();
+        });
+
 
         PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerSaveButton.addActionListener(e -> {
-            ConsoleLoggerSettings.setPattern(0, propertiesConsoleLoggerTextField1.getText());
-            ConsoleLoggerSettings.setPattern(1, propertiesConsoleLoggerTextField2.getText());
-            ConsoleLoggerSettings.setPattern(2, propertiesConsoleLoggerTextField3.getText());
-            ConsoleLoggerSettings.setPattern(3, propertiesConsoleLoggerTextField4.getText());
-            ConsoleLoggerSettings.setPattern(4, propertiesConsoleLoggerTextField5.getText());
-            ConsoleLoggerSettings.setPattern(5, propertiesConsoleLoggerTextField6.getText());
-            ConsoleLoggerSettings.setPattern(6, propertiesConsoleLoggerTextField7.getText());
-            ConsoleLoggerSettings.setPattern(7, propertiesConsoleLoggerTextField8.getText());
-            ConsoleLoggerSettings.setPattern(8, propertiesConsoleLoggerTextField9.getText());
+            if (preset == 1) {
+                saveActiveLoggers();
+                save1Loggers();
+            } else      if (preset == 2) {
+                saveActiveLoggers();
+                save2Loggers();
+            }
         });
+
     }
 
-    private void loadAllLoggers() {
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField1.setText(ConsoleLoggerSettings.getPattern(0));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField2.setText(ConsoleLoggerSettings.getPattern(1));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField3.setText(ConsoleLoggerSettings.getPattern(2));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField4.setText(ConsoleLoggerSettings.getPattern(3));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField5.setText(ConsoleLoggerSettings.getPattern(4));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField6.setText(ConsoleLoggerSettings.getPattern(5));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField7.setText(ConsoleLoggerSettings.getPattern(6));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField8.setText(ConsoleLoggerSettings.getPattern(7));
-        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField9.setText(ConsoleLoggerSettings.getPattern(8));
+    private void saveActiveLoggers() {
+        ConsoleLoggerSettings.setPattern(0, propertiesConsoleLoggerTextField1.getText());
+        ConsoleLoggerSettings.setPattern(1, propertiesConsoleLoggerTextField2.getText());
+        ConsoleLoggerSettings.setPattern(2, propertiesConsoleLoggerTextField3.getText());
+        ConsoleLoggerSettings.setPattern(3, propertiesConsoleLoggerTextField4.getText());
+        ConsoleLoggerSettings.setPattern(4, propertiesConsoleLoggerTextField5.getText());
+        ConsoleLoggerSettings.setPattern(5, propertiesConsoleLoggerTextField6.getText());
+        ConsoleLoggerSettings.setPattern(6, propertiesConsoleLoggerTextField7.getText());
+        ConsoleLoggerSettings.setPattern(7, propertiesConsoleLoggerTextField8.getText());
+        ConsoleLoggerSettings.setPattern(8, propertiesConsoleLoggerTextField9.getText());
+    }
+
+    private void save1Loggers() {
+        ConsoleLoggerSettings.setPattern(9, propertiesConsoleLoggerTextField1.getText());
+        ConsoleLoggerSettings.setPattern(10, propertiesConsoleLoggerTextField2.getText());
+        ConsoleLoggerSettings.setPattern(11, propertiesConsoleLoggerTextField3.getText());
+        ConsoleLoggerSettings.setPattern(12, propertiesConsoleLoggerTextField4.getText());
+        ConsoleLoggerSettings.setPattern(13, propertiesConsoleLoggerTextField5.getText());
+        ConsoleLoggerSettings.setPattern(14, propertiesConsoleLoggerTextField6.getText());
+        ConsoleLoggerSettings.setPattern(15, propertiesConsoleLoggerTextField7.getText());
+        ConsoleLoggerSettings.setPattern(16, propertiesConsoleLoggerTextField8.getText());
+        ConsoleLoggerSettings.setPattern(17, propertiesConsoleLoggerTextField9.getText());
+    }
+
+    private void save2Loggers() {
+        ConsoleLoggerSettings.setPattern(18, propertiesConsoleLoggerTextField1.getText());
+        ConsoleLoggerSettings.setPattern(19, propertiesConsoleLoggerTextField2.getText());
+        ConsoleLoggerSettings.setPattern(20, propertiesConsoleLoggerTextField3.getText());
+        ConsoleLoggerSettings.setPattern(21, propertiesConsoleLoggerTextField4.getText());
+        ConsoleLoggerSettings.setPattern(22, propertiesConsoleLoggerTextField5.getText());
+        ConsoleLoggerSettings.setPattern(23, propertiesConsoleLoggerTextField6.getText());
+        ConsoleLoggerSettings.setPattern(24, propertiesConsoleLoggerTextField7.getText());
+        ConsoleLoggerSettings.setPattern(25, propertiesConsoleLoggerTextField8.getText());
+        ConsoleLoggerSettings.setPattern(26, propertiesConsoleLoggerTextField9.getText());
+    }
+
+    private void load1Loggers() {
+        preset = 1;
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField1.setText(ConsoleLoggerSettings.getPattern(9));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField2.setText(ConsoleLoggerSettings.getPattern(10));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField3.setText(ConsoleLoggerSettings.getPattern(11));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField4.setText(ConsoleLoggerSettings.getPattern(12));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField5.setText(ConsoleLoggerSettings.getPattern(13));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField6.setText(ConsoleLoggerSettings.getPattern(14));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField7.setText(ConsoleLoggerSettings.getPattern(15));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField8.setText(ConsoleLoggerSettings.getPattern(16));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField9.setText(ConsoleLoggerSettings.getPattern(17));
+    }
+
+    private void load2Loggers() {
+        preset = 2;
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField1.setText(ConsoleLoggerSettings.getPattern(18));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField2.setText(ConsoleLoggerSettings.getPattern(19));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField3.setText(ConsoleLoggerSettings.getPattern(20));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField4.setText(ConsoleLoggerSettings.getPattern(21));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField5.setText(ConsoleLoggerSettings.getPattern(22));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField6.setText(ConsoleLoggerSettings.getPattern(23));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField7.setText(ConsoleLoggerSettings.getPattern(24));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField8.setText(ConsoleLoggerSettings.getPattern(25));
+        PropertiesConsoleLoggerToolSetup.propertiesConsoleLoggerTextField9.setText(ConsoleLoggerSettings.getPattern(26));
     }
 
     @Override