+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.0.29] - 13-09-2024
 - changed some of the input field sizes
+- added option for second preset of loggers
 
 ## [0.0.28] - 06-09-2024
 - Saving seems to work better now